day10: add caching
Turns out the input is long, but repeats digits, so by processing digits
individually, I win.

To make it perfect, I'd use multicore ocaml to process the different
inputs (in that case I'd probably create a set from the input string, as
the cache wouldn't be as effective). The problem is that multicore ocaml
doesn't work on arm64 yet.